mauritania-chatbot/requirements.txt
gradio>=4.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
groq>=0.3.0
python-dotenv>=1.0.0